+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.3.0] - 2022-10-20
+
 ### Changed
 
 - Add `mc_sgx_tservice::SealError`, make `SealedBuilder` use it instead of
@@ -64,7 +66,8 @@ doc builds for:
 - Initial release of FFI bindings crates
 
 <!-- next-url -->
-[Unreleased]: https://github.com/mobilecoinfoundation/sgx/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/mobilecoinfoundation/sgx/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/mobilecoinfoundation/sgx/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/mobilecoinfoundation/sgx/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/mobilecoinfoundation/sgx/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/mobilecoinfoundation/sgx/compare/v0.1.0